+++ src/core/enhancers.ts
@@ -7,6 +7,7 @@ import {
 } from './helpers';
 import type { SigilOptions } from './options';
 import type { TypedSigil } from './types';
+import { __DEV__ } from './constants';
 
 /**
  * HOF (class inhancer) that attaches runtime sigil metadata to Sigil class.
@@ -46,7 +47,7 @@ export function withSigil<S extends Function, L extends string = string>(
   // decorate and check inheritance.
   const ctor = Class;
   decorateCtor(ctor, l);
-  checkInheritance(ctor, opts);
+  if (__DEV__) checkInheritance(ctor, opts);
 
   return Class;
 }
@@ -90,7 +91,7 @@ export function withSigilTyped<S extends Function, L extends string = string>(
   // decorate and check inheritance.
   const ctor = Class;
   decorateCtor(ctor, l);
-  checkInheritance(ctor, opts);
+  if (__DEV__) checkInheritance(ctor, opts);
 
   return Class as unknown as TypedSigil<S, L>;
 }